RAG_law/requirements.txt
streamlit
langchain-core
langchain-community
langchain-openai
langchain-ollama
faiss-cpu
python-dotenv
tavily-python
unstructured
pypdf
docx2txt

